requirements.txt
AC-Flask-HipChat==0.2.2
Flask==0.10.1
Flask-OAuthlib==0.7.0
Jinja2==2.7.3
MarkupSafe==0.23
Padding==0.4
PyJWT==0.3.0
Werkzeug==0.9.6
basicauth==0.2
boto==2.34.0
itsdangerous==0.24
keen==0.3.6
newrelic==2.40.0.34
oauthlib==0.7.2
pycrypto==2.6.1
pymongo==2.7.2
redis==2.10.3
requests==2.4.3
wsgiref==0.1.2
Flask-RESTful==0.3.1
mock==1.0.1
